Move static files to assets/
