+++ client/src/assets/scoreDescriptors.js
