--- client/src/components/Result/Profile/profiles.js
@@ -1,53 +1,0 @@
-const agileProfiles = [
-  {
-    name: 'Dhanaraj Jayapalan',
-    position: 'Product Owner',
-    department: 'Singapore',
-    email: 'dhanaraj.jayapalan@zuhlke.com',
-    phone: '+65 92231019',
-    pic: './raj.png',
-    linkedin: 'https://www.linkedin.com/in/dhanarajjayapalan/',
-  },
-  {
-    name: 'Aleksandra Brewer',
-    position: 'Project Manager',
-    department: 'Singapore',
-    email: 'aleksandra.brewer@zuhlke.com',
-    pic: './alex.png',
-    linkedin: 'https://www.linkedin.com/in/aleksandra-brewer-7020594/',
-  },
-  {
-    name: 'Jochen Reber',
-    position: 'Principal Consultant',
-    department: 'Singapore',
-    email: 'jochen.reber@zuhlke.com',
-    phone: '+65 8459 3220',
-    pic: './jochen.jpg',
-    linkedin: 'https://www.linkedin.com/in/jochen-reber-7576a3181/',
-  },
-  {
-    name: 'Michael Schnyder',
-    position: 'Agile Coach / Lead Software Architect',
-    department: 'Singapore',
-    email: 'michael.schnyder@zuhlke.com',
-    phone: '+65 8182 5581',
-    pic: './michael.jpg',
-    linkedin: 'https://www.linkedin.com/in/mschnyder/',
-  },
-];
-
-const salesProfiles = [
-  {
-    name: 'Kiky Shannon',
-    position: 'Business Development Manager',
-    department: 'Singapore',
-    email: 'kiky.shannon@zuhlke.com',
-    phone: '+65 8268 8116',
-    pic: './kiky.jpeg',
-    linkedin: 'https://www.linkedin.com/in/kikyshannon/',
-  },
-];
-
-const displayedProfiles = [agileProfiles[2], salesProfiles[0]];
-
-export default displayedProfiles;
